# ai-service/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "江苏 13 市 AI 稽核助手",
    "description": "江苏 13 市 AI 稽核助手 REST API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analysis-runs/{run_id}/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Create Report",
        "description": "把完成的AI草稿保存为可编辑报告；重复调用返回同一资源。",
        "operationId": "create_report_api_v1_analysis_runs__run_id__reports_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report",
        "description": "读取报告当前版本、状态和版本历史。",
        "operationId": "get_report_api_v1_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "reports"
        ],
        "summary": "Update Report",
        "description": "保存一个新的完整报告版本。",
        "operationId": "update_report_api_v1_reports__report_id__patch",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/reviews": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Review Report",
        "description": "提交、退回或审核通过一份报告。",
        "operationId": "review_report_api_v1_reports__report_id__reviews_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportReview"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/content": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Report",
        "description": "仅下载当前城市已经审核通过的正式Word。",
        "operationId": "download_report_api_v1_reports__report_id__content_get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sites": {
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "新增当前城市报账点",
        "description": "新增独立报账点主数据；不会同时生成材料、案例或纠错记忆。",
        "operationId": "create_site_api_v1_sites_post",
        "parameters": [
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "查询当前城市报账点",
        "description": "支持前端远程搜索，避免一次加载全部报账点。",
        "operationId": "list_sites_api_v1_sites_get",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportReview": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "submit",
              "return",
              "approve"
            ],
            "title": "Action"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ReportReview",
        "description": "审核状态动作；退回时必须说明原因。"
      },
...
      "ReportUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Title"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSectionEdit"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 6,
            "title": "Sections"
          },
          "change_summary": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Change Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "sections",
          "change_summary"
        ],
        "title": "ReportUpdate",
        "description": "保存一个完整的新版本，避免局部更新留下章节缺失。"
      },
      "ReportVersionSummary": {
        "properties": {
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "has_docx": {
            "type": "boolean",
            "title": "Has Docx"
          }
        },
        "type": "object",
        "required": [
          "version_no",
          "change_summary",
          "created_by",
          "created_at",
          "has_docx"
        ],
        "title": "ReportVersionSummary"
      },
      "ReportVersionView": {
        "properties": {
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSectionView"
            },
            "type": "array",
            "title": "Sections"
          },
          "uncertain_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertain Items"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary"
          },
          "has_docx": {
            "type": "boolean",
            "title": "Has Docx"
          },
          "docx_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docx Sha256"
          },
          "docx_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docx Size Bytes"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "version_no",
          "title",
          "sections",
          "uncertain_items",
          "review_reasons",
          "change_summary",
          "has_docx",
          "docx_sha256",
          "docx_size_bytes",
          "created_by",
          "created_at",
          "generated_at"
        ],
        "title": "ReportVersionView"
      },
      "ReportView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "city_code": {
            "type": "string",
            "title": "City Code"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "analysis_run_id": {
            "type": "string",
            "title": "Analysis Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "in_review",
              "returned",
              "approved"
            ],
            "title": "Status"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "approved_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Version"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "current": {
            "$ref": "#/components/schemas/ReportVersionView"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ReportVersionSummary"
            },
            "type": "array",
            "title": "Versions"
          },
          "content_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Url"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city_code",
          "site_id",
          "task_id",
          "analysis_run_id",
          "status",
          "current_version",
          "approved_version",
          "review_note",
          "current",
          "versions",
          "content_url",
          "created_by",
          "reviewed_by",
          "created_at",
          "updated_at",
          "reviewed_at"
        ],
        "title": "ReportView"
      },
      "SiteCreate": {
        "properties": {
          "site_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Site Name",
            "description": "报账点正式名称；用于页面展示和历史材料人工关联。"
          },
          "site_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Code",
            "description": "稳定业务编码；为空表示由系统生成内部编码。"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Address",
            "description": "报账点地址；未知时为空，不参与城市隔离和记忆检索。"
          }
        },
        "type": "object",
        "required": [
          "site_name"
        ],
        "title": "SiteCreate",
        "description": "创建报账点时由页面提交的主数据字段。\n\n城市不允许出现在请求体中，而是由经过后端校验的 ``X-City-Code`` 上下文决定，\n这样即使前端数据过期，也不能把苏州报账点写入南京。编码为空时由后端生成稳定\n的内部编码；对接资管系统后可以直接提交真实业务编码。"
      },
      "SiteList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SiteView"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "SiteList",
        "description": "报账点分页结果。"
      },
      "SiteView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "city_code": {
            "type": "string",
            "title": "City Code"
          },
          "site_code": {
            "type": "string",
            "title": "Site Code"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city_code",
          "site_code",
          "site_name",
          "address",
          "status"
        ],
        "title": "SiteView",
        "description": "前端下拉框和列表所需的报账点字段。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}